Added Necessary Views & Added Styles
I added the necessary views, and gave them each a title, which briefly describes what that page entails.

I also added styles to the site.css, to help keep the flow of the page ordered and neat. These containers will ensure that, in the case when other functionality is to be added to the Web App, it will not be difficult to manage and organize these additional features on the pages.