Updated Register View
Updated Register.cshtml to model the Lecturer model, with the form submissions.

Can now successfully call the action RegisterUser in the HomeController which checks to see if the model state is valid, and if it is, it adds the model to the appropriate table and saves the database with the newly inserted data.

Added sample data to the ProgrammeCoordinator table and AcademicManager table to test for login feature later.