Successfully Created Submit-Claim Functionality & Updated Claim Model
Updated the Claim model to store the data of the file in a byte array.

Added migrations to add foreign keys to necessary tables.

Updated HomeController method SubmitNewClaim() which takes to parameters, one for the Claim model and the other is a IFormFile which contains the data of the file that is uploaded.

Updated database after migrations.