Updated HomeController to Review Claims & ReviewClaim View
Removed previous hard-coded sample data from tables in ReviewClaim.cshtml and adjust the table rows accordingly.

Specifying the expected model type for the view so it can expect a collection of Claim objects, allowing for enumeration.

Added methods to the HomeController, allowing the user to download the file relating to the Claim. Methods were also added to approve and reject claim requests.

When the ReviewClaim is called and it returns the View() a list of Claims that includes the lecturer attributes is included, to further display their name in the claim request table.

Had to change the access level of the DbSet<Claim> in the AppDbContext.cs to public to gain further access in the HomeController.