Added File Type Validation
Added a file type validation that can only be 1 of the following: .pdf / .docx / .png / .jpg.

A string array storing the various valid file types is created in the method, only if a file has been uploaded and is not null.

Then a if statement iterates through the array of valid file types, comparing the end of the uploaded file name's string, ignoring case sensitivity. If it exists, it returns true, else it returns false. If it is false, it returns the view CreateClaim with the TempData notifying the user of the incorrect file type.