Refactored database structure to simplify login and access control.
Unified the [User] table to merge program coordinators, academic managers, and lecturers for streamlined authentication and role-based authorization.

Added a migration to recreate the tables in the local database after encountering issues with table renaming, which caused build errors.

Deleted the database and recreated it to ensure the new schema works and can handle rollbacks to previous versions.

Updated the schema and added a migration to drop the previous tables whilst I modify the structure of my database.