Updated Views, Database Schema and Controller
I was experiencing problems with adding migrations and updating the database. The new changes were not being detected, and therefore no modifications were being applied.

Deleted all previous migrations, and re-created the database for the project with the updated database schema.

Updated Views and Controllers, to accommodate for new database schema.