Added Auto-Calculator & Format
Added a auto-calculate function using Jquery in the form where lecturers can submit claims. It is declared as read-only so the user may not be able to change the value.

Added a Validate User Access Level, to only allow lecturers to submit claims.

Document: Formatted the Home Controller for easier readability.